+++ spec/lib/karafka/web/ui/lib/paginations/paginators/partitions_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  subject(:pagination) { described_class.call(partitions_count, page) }
+
+  pending
+end